ayaka ascension refactor
