+++ internal/characters/ayaka/skill.go
@@ -5,8 +5,6 @@ import (
 	"github.com/genshinsim/gcsim/pkg/core/action"
 	"github.com/genshinsim/gcsim/pkg/core/attributes"
 	"github.com/genshinsim/gcsim/pkg/core/combat"
-	"github.com/genshinsim/gcsim/pkg/core/player/character"
-	"github.com/genshinsim/gcsim/pkg/modifier"
 )
 
 var skillFrames []int
@@ -34,16 +32,7 @@ func (c *char) Skill(p map[string]int) action.ActionInfo {
 		Mult:       skill[c.TalentLvlSkill()],
 	}
 
-	//a1 increase normal + ca dmg by 30% for 6s
-	m := make([]float64, attributes.EndStatType)
-	m[attributes.DmgP] = 0.3
-	c.AddAttackMod(character.AttackMod{
-		Base: modifier.NewBaseWithHitlag("ayaka-a1", 360),
-		Amount: func(atk *combat.AttackEvent, t combat.Target) ([]float64, bool) {
-			return m, atk.Info.AttackTag == combat.AttackTagNormal || atk.Info.AttackTag == combat.AttackTagExtra
-		},
-	})
-
+	c.a1()
 	c.Core.QueueAttack(ai, combat.NewCircleHitOnTarget(c.Core.Combat.Player(), nil, 4.5), 0, skillHitmark)
 
 	// 4 or 5, 1:1 ratio
